Database finished 2
The database is finished and saves to a json file

i rename the menu files to make more consistant

i renamed the class names to be more consisytant

remove where it asks for your name in. some files. it now reads form database

changed some menus to be more user friendly